Miscellaneous/percyGPT/src/requirements.txt
flask
openai
python-dotenv
Flask-Limiter
gunicorn
